Move a `SecretStorage` unit test into its own file
